+++ spinner/spinner_windows.go
@@ -1,0 +1,14 @@
+//go:build windows
+
+package spinner
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+//                                                                                    //
+//                         Copyright (c) 2026 ESSENTIAL KAOS                          //
+//      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
+//                                                                                    //
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+func getMaxDescSize() int {
+	return 120
+}