+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [14.1.9](https://kaos.sh/ek/14.1.9)
 
 - **`[terminal]`** Fixed bug with reading data on Windows
+- **`[spinner]`** Fixed Windows support
 
 ### [14.1.8](https://kaos.sh/ek/14.1.8)